refactor: 패키지 구조 변경
rules 디렉토리 생성
