--- app/src/main/java/woowacourse/movie/domain/SeatsFactory.kt
@@ -1,5 +1,0 @@
-package woowacourse.movie.domain
-
-interface SeatsFactory {
-    fun get(): List<Seat>
-}
