+++ app/src/main/java/woowacourse/movie/domain/rules/RunningTimeRule.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.rules
 
 import java.time.DayOfWeek
 import java.time.LocalDate
