+++ app/src/main/java/woowacourse/movie/domain/rules/SeatsFactory.kt
@@ -1,0 +1,7 @@
+package woowacourse.movie.domain.rules
+
+import woowacourse.movie.domain.Seat
+
+interface SeatsFactory {
+    fun get(): List<Seat>
+}
