+++ app/src/test/java/woowacourse/movie/domain/RunningTimeRuleTest.kt
@@ -3,6 +3,8 @@ package woowacourse.movie.domain
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import woowacourse.movie.domain.rules.RunningTimeRule
+import woowacourse.movie.domain.rules.RunningTimeRuleImpl
 import woowacourse.movie.fixture.DomainFixture
 import java.time.LocalDate
 import java.time.LocalDateTime
